DDF-04647 Fix bug preventing users from changing search forms
